refactor: schedule benchmarks
